chore - #1501 move codegraph record production behind the compiler boundary
`cli/commands/codegraph.rs` held 3,344 lines of compiler analysis in a
module whose job is a CLI command. RFC 120 names the failure mode that
invites -- "a check living in the CLI could not agree with one living in
the frontend by construction" -- and #1293 found a live instance of it in
build.rs. This is the same class, a different file, and it was not in the
recorded reach-in set.

The seam was already clean: inspect_codegraph normalized a path, called
collect_codegraph_records, and formatted the result. So this is a
relocation, not a rewrite. The file stays one file; splitting it is a
mechanical sweep worth doing once the code has settled in its new home,
where the boundaries will be evident rather than guessed.

Production lands in src/inspect/, named for compiler/incan_inspect, which
is where loaves/LAYOUT.md sends rust_inspect. The command surface stays
in the CLI and goes to toolchain/incan-cli. RFC 118 puts codegraph on the
incan surface rather than oven, and names the alternative directly:
"`oven inspect` must not masquerade as a semantic codegraph API".

Two supporting moves were forced rather than chosen. ParsedModule is the
product of parsing and was defined in cli::prelude, so a compiler-side
module would have had to import from crate::cli to name its own input; it
moves to the frontend and cli::prelude re-exports it, leaving every other
caller untouched. And the analysis now owns CodegraphError instead of
borrowing CliError, with conversion at the command boundary.

One edge remains and it is inherited, not new. The producer calls
common.rs for session and discovery, which LAYOUT.md row 76 already sends
to compiler/incan_driver, and which oven/legacy_cargo.rs, lsp/backend.rs
and backend/project/cargo_toml.rs reach the same way today. Cutting it
means extracting incan_driver from an 8,691-line module with ten
dependents (#1479), not editing this file; when that lands, the edge
resolves here with no further work.

Adding From<CodegraphError> for CliError made one closure in build.rs
ambiguous between two conversions. Annotated; no behaviour changes.

Verified as a move: the only content removed from the relocated file is
the two CLI imports, the nineteen-line entry point, the CliResult and
CliError spellings, and three visibility keywords. No analysis was lost.
Codegraph tests pass and the named lib-suite failure set is unchanged at
the 30 this branch already carried.
