+++ src/lib.rs
@@ -23,6 +23,7 @@ pub mod format;
 pub mod frontend;
 #[cfg(feature = "cli")]
 pub(crate) mod generated_cache;
+pub mod inspect;
 pub mod library_manifest;
 pub mod lockfile;
 #[cfg(feature = "lsp")]
